requirements.txt
annotated-types==0.7.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
python-dotenv==1.0.1
rsa==4.2
Telethon==1.39.0
typing_extensions==4.12.2
